+++ packages/web-sdk/src/index.ts
@@ -167,5 +167,4 @@ export {
   STORAGE_KEY,
 } from './instrumentations/session';
 
-export type { ErrorDetails } from './utils/errors';
 export { getIgnoreUrls } from './utils/url';
