+++ packages/web-sdk/src/instrumentations/errors/getErrorDetails.test.ts
@@ -1,17 +1,23 @@
 import {
   defaultErrorArgsSerializer,
   type ExceptionEvent,
+  initializeFaro,
   LogEvent,
   stringifyExternalJson,
   type TransportItem,
 } from '@grafana/faro-core';
 import { mockConfig, MockTransport } from '@grafana/faro-core/src/testUtils';
 
-import { initializeFaro } from '../initialize';
-import { ConsoleInstrumentation } from '../instrumentations';
-import { registerOnerror } from '../instrumentations/errors/registerOnerror';
+import { ConsoleInstrumentation } from '../console';
 
-import { ErrorDetails, getDetailsFromConsoleErrorArgs, getDetailsFromErrorArgs } from './errors';
+import { ErrorDetails, getDetailsFromConsoleErrorArgs, getDetailsFromErrorArgs } from './getErrorDetails';
+import { registerOnerror } from './registerOnerror';
+
+// import { initializeFaro } from '../initialize';
+// import { ConsoleInstrumentation } from '../instrumentations';
+// import { registerOnerror } from '../instrumentations/errors/registerOnerror';
+
+// import { ErrorDetails, getDetailsFromConsoleErrorArgs, getDetailsFromErrorArgs } from './errors';
 
 describe('errors', () => {
   it('parses an Error passed to console.error', () => {
